misc(moneyhash): Refact webhook handling
